CIV-3107 moved final document form to model
